refactor: move css imports to the file that uses them
Instead of importing in main.
