+++ log-viewer/modules/calltree-view/CalltreeView.ts
@@ -7,10 +7,9 @@
 //todo: add filter on log level (fine, finer etc)
 import { RowComponent, TabulatorFull as Tabulator } from 'tabulator-tables';
 
-import '../../resources/css/DatabaseView.scss';
-import '../../resources/css/TreeView.css';
 import { rootMethod } from '../Main';
 import { showTab } from '../Util';
+import '../database-view/DatabaseView.scss';
 import MinMaxEditor from '../datagrid/editors/MinMax';
 import MinMaxFilter from '../datagrid/filters/MinMax';
 import NumberFormat from '../datagrid/format/Number';
@@ -18,6 +17,7 @@ import { RowKeyboardNavigation } from '../datagrid/module/RowKeyboardNavigation'
 import { RowNavigation } from '../datagrid/module/RowNavigation';
 import { LogLine, RootNode, TimedNode } from '../parsers/TreeParser';
 import { hostService } from '../services/VSCodeService';
+import './TreeView.css';
 
 let calltreeTable: Tabulator;
